+++ Editor/Package.cs
@@ -7,7 +7,7 @@ namespace HananokiEditor.SymbolSettings {
     public const string name = "SymbolSettings";
     public const string nameNicify = "Symbol Settings";
     public const string editorPrefName = "Hananoki.SymbolSettings";
-    public const string version = "1.0.8";
+    public const string version = "2.0.0";
     public static string projectSettingsPath => $"{SharedModule.SettingsEditor.projectSettingDirectory}/SymbolSettings.json";
   }
 }